chore: hero makeover and bug fixes
