+++ packages/website/components/hero/index.js
@@ -1,0 +1,2 @@
+import Hero from './hero'
+export default Hero
