+++ packages/website/styles/global.css
@@ -69,8 +69,6 @@
 }
 
 .typography-hero-subtitle {
-  font-family: "Open Sans", sans-serif;
-  font-size: 1.5rem;
   letter-spacing: 0.02em;
 }
